moved all methods at once
